drm/amdkfd: prepare debug event status for queue suspend
Queue suspend will require an update of debug event status.
READ/WRITE_ONCE will be used to avoid race condition.

+++ drivers/gpu/drm/amd/amdkfd/kfd_priv.h
@@ -533,7 +533,7 @@ struct queue_properties {
 	/* Relevant for CU */
 	uint32_t cu_mask_count; /* Must be a multiple of 32 */
 	uint32_t *cu_mask;
-	unsigned int debug_event_type;
+	unsigned long debug_event_type;
 };
 
 #define QUEUE_IS_ACTIVE(q) ((q).queue_size > 0 &&	\
